Finished Splitting the DBConnection, fixed small Bugs concerning Buttons not working, Finished Splitting the Tests by domain. - Pair Programming
